Extract DumpContext base class from MinidumpContext
This is a pure refactoring in preparation of the Microdump(s)
(which will be introduced by later CL). No behavioral change is
intended.
This change removes the MinidumpContext -> MinidumpStream
inheritance (which is not actually needed as MinidumpContexts
are never stored to streams) and extracts a base class
(DumpContext) out of it.
Similarly, ProcessResult is moved out to its own class.

BUG=chromium:410294

Review URL: https://breakpad.appspot.com/5684002/